# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kang Raling API Web Service",
...
        "contact": {
            "name": "Developer Website Kang Raling",
            "email": "danu3anugrah@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server Lokal Development"
        },
        {
            "url": "https://kangraling.dlhgarut.id",
            "description": "Server Utama (Production Dinas Lingkungan Hidup)"
        }
    ],
    "paths": {
...
        "/edukasis": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil semua konten edukasi",
                "description": "Mengembalikan daftar konten edukasi untuk halaman publik. Mendukung filter berdasarkan kategori (`modul` atau `video`). Tidak membutuhkan token.",
                "operationId": "1d2349298323d1031eb8e1b92c822f38",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter kategori konten",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "modul",
                                "video"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar konten edukasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Modul Pemilahan Sampah"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "modul-pemilahan-sampah"
                                                    },
                                                    "kategori": {
                                                        "type": "string",
                                                        "example": "modul",
                                                        "enum": [
                                                            "modul",
                                                            "video"
                                                        ]
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "file_pdf": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "link_video": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "gambar": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Edukasi"
                ],
                "summary": "Tambah konten edukasi baru",
                "description": "Menambahkan konten edukasi baru.\n\n**Aturan kategori:**\n- Kategori `modul` → wajib upload `file_pdf`\n- Kategori `video` → wajib isi `link_video` (URL YouTube)\n\nGunakan `multipart/form-data` karena ada upload file.",
                "operationId": "a4aee92801de3649c719b75776788d22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "kategori"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Modul Pemilahan Sampah Organik"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "modul",
                                        "enum": [
                                            "modul",
                                            "video"
                                        ]
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Panduan lengkap pemilahan sampah organik",
                                        "nullable": true
                                    },
                                    "file_pdf": {
                                        "description": "Wajib jika kategori = modul",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "link_video": {
                                        "description": "Wajib jika kategori = video",
                                        "type": "string",
                                        "example": "https://youtube.com/watch?v=xxx",
                                        "nullable": true
                                    },
                                    "gambar": {
                                        "description": "Gambar thumbnail",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Konten edukasi berhasil ditambahkan"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.edukasi"
                    },
                    "422": {
                        "description": "Validasi gagal — PDF tidak diupload untuk modul / link video kosong untuk video"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/edukasis/{slug}": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil detail konten edukasi berdasarkan slug",
                "description": "Mengembalikan detail lengkap satu konten edukasi. Tidak membutuhkan token.",
                "operationId": "3be245e454f6f2418cdc06611260606c",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "modul-pemilahan-sampah"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail konten edukasi berhasil diambil"
                    },
                    "404": {
                        "description": "Konten edukasi tidak ditemukan"
                    }
                }
            }
        },
        "/edukasis/{id}/edit": {
            "get": {
                "tags": [
                    "Edukasi"
                ],
                "summary": "Ambil data edukasi untuk form edit (dashboard)",
                "description": "Mengambil data konten edukasi berdasarkan ID untuk form edit di dashboard. Fasilitator hanya bisa akses konten miliknya sendiri.",
                "operationId": "2c503c9c9b3deea8ef0fb0279c0a2e4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data edukasi untuk edit berhasil diambil"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Konten edukasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/edukasis/{id}": {
            "put": {
                "tags": [
                    "Edukasi"
                ],
                "summary": "Update konten edukasi",
                "description": "Mengupdate konten edukasi yang sudah ada. Gunakan `multipart/form-data` dengan `_method: PUT`.",
                "operationId": "9e1be3240c4ebfa3f691cd3bcb46ba0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "kategori"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Modul Pemilahan Sampah Organik (Revisi)"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "modul",
                                        "enum": [
                                            "modul",
                                            "video"
                                        ]
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "file_pdf": {
                                        "description": "Kosongkan jika tidak ganti PDF",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "link_video": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konten edukasi berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Konten edukasi tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Edukasi"
                ],
                "summary": "Hapus konten edukasi",
                "description": "Menghapus konten edukasi secara permanen. Fasilitator hanya bisa hapus konten miliknya sendiri.",
                "operationId": "e5b9562e91093ffa8af76bb8bdb342a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konten edukasi berhasil dihapus"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Konten edukasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dengan format: **Bearer {token}**",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}